Removed some unneccessary casts.

+++ HTTPHelper C-Sharp/HTTPHelper.cs
@@ -867,12 +867,12 @@ public class HTTPHelper
             CaptureSSLInfo(fileDownloadURL, ref httpWebRequest);
 
             // Gets the size of the remote file on the web server.
-            remoteFileSize = (long)webResponse.ContentLength;
+            remoteFileSize = webResponse.ContentLength;
 
             Stream responseStream = webResponse.GetResponseStream();
             // Gets the response stream.
 
-            long lngBytesReadFromInternet = (long)responseStream.Read(dataBuffer, 0, dataBuffer.Length);
+            long lngBytesReadFromInternet = responseStream.Read(dataBuffer, 0, dataBuffer.Length);
             // Reads some data from the HTTP stream into our data buffer.
 
             // We keep looping until all of the data has been downloaded.
@@ -890,7 +890,7 @@ public class HTTPHelper
                 // Update the download percentage value.
                 DownloadStatusUpdateInvoker();
 
-                lngBytesReadFromInternet = (long)responseStream.Read(dataBuffer, 0, dataBuffer.Length);
+                lngBytesReadFromInternet = responseStream.Read(dataBuffer, 0, dataBuffer.Length);
                 // Reads more data into our data buffer.
             }
 
@@ -1001,14 +1001,14 @@ public class HTTPHelper
             CaptureSSLInfo(fileDownloadURL, ref httpWebRequest);
 
             // Gets the size of the remote file on the web server.
-            remoteFileSize = (long)webResponse.ContentLength;
+            remoteFileSize = webResponse.ContentLength;
 
             Stream responseStream = webResponse.GetResponseStream();
             // Gets the response stream.
             fileWriteStream = new FileStream(localFileName, FileMode.Create);
             // Creates a file write stream.
 
-            long lngBytesReadFromInternet = (long)responseStream.Read(dataBuffer, 0, dataBuffer.Length);
+            long lngBytesReadFromInternet = responseStream.Read(dataBuffer, 0, dataBuffer.Length);
             // Reads some data from the HTTP stream into our data buffer.
 
             // We keep looping until all of the data has been downloaded.
@@ -1026,7 +1026,7 @@ public class HTTPHelper
                 // Update the download percentage value.
                 DownloadStatusUpdateInvoker();
 
-                lngBytesReadFromInternet = (long)responseStream.Read(dataBuffer, 0, dataBuffer.Length);
+                lngBytesReadFromInternet = responseStream.Read(dataBuffer, 0, dataBuffer.Length);
                 // Reads more data into our data buffer.
             }
